common: A set of useful functions

The common package is set of useful functions that enhance base R. These functions make up for some deficiencies in the base package. For example, there are functions to sort a data frame by multiple columns, and manipulate data frame labels. These are common activities, and can be used in almost any R program.

Functions Included

The functions included in the common package are as follows:

v: A generalized NSE quoting function.

sort.data.frame: An overload of the sort function for data frames.

labels.data.frame: An overload of the labels function for data frames.

%p%: An infix operator for paste0().

%eq%: An enhanced equality infix operator.

roundup: A rounding function that matches SAS® software.

Sys.path: A function to return the path of the current program.

file.find: A function to search for a file on the file system.

dir.find: A function to search for directories on the file system.

find.names: A function to search for variable names on a data frame.

copy.attributes: A function to copy column attributes from one data frame to another.

supsc: A function to get UTF-8 superscript characters.

subsc: A function to get UTF-8 subscript characters.