Search for files

A function to find files on the file system. The function starts from the directory specified in the location parameter, and searches for the file name in the file_name parameter. The file_name accepts wild cards (*). The function searches both above and below the current directory locations. The number of levels to search can be controlled by the up and down parameters.

file.find(location, file_name, up = 3, down = 3)

Arguments

location
The directory to start searching from. The path of the current program can be retrieved using Sys.path.

file_name
A full or partial name of the file to find. If partial, use the asterisk (*) character to indicate the missing piece.

up
The number of levels above the location to search.

down
The number of levels below the location to search.

Value

A vector of one or more full file paths that met the search criteria, or a NULL if no files met the criteria.

Examples

# Search for a file named "globals.R"
res <- file.find(Sys.path(), "globals.R")

# Search for SAS datasets
res <- file.find(Sys.path(), "*.sas7bdat")