Gets UTF-8 symbol characters

The symbol function gets UTF-8 symbol characters. You can call this function to look up trademarks, Greek letters, and mathematical operators. The function uses HTML entity keywords to indicate which symbol to return. You may pass more than one keyword in a single call to get a combined result. Any characters not recognized as a keyword will be left alone. Characters surrounded by square brackets ([]) will be subscripted. Characters surrounded by square brackets and prefixed with an up arrow (^[]) will be superscripted.

symbol(keyword)

Arguments

keyword
A symbol keyword. This keyword follows HTML conventions. See the Keywords section for a complete list of all supported keywords.

Value

The requested UTF-8 symbol character or characters.

Keywords

The following symbol keywords are available:

Trademark and Copyright: copy, reg, trade

Financial: cent, euro, pound, rupee, ruble, yen, yuan

Mathmatical: asymp, bcong, cong, coprod, empty, fnof, ge, int, Int, le, ncong, ne, not, part, plusmn, prod, radic, sime, sum

Logical: and, cap, cup, comp, cuvee, cuwed, exist, forall, fork, isin, nexist, ni, notin, notni, nsub, nsup, or, sub, sup, xcap, xcup, xvee, xwedge

Greek uppercase letters: Alpha, Beta, Gamma, Delta, Epsilon, Zeta, Eta, Theta, Iota, Kappa, Lambda, Mu, Nu, Xi, Omicron, Pi, Rho, Sigma, Tau, Upsilon, Phi, Chi, Psi, Omega

Greek lowercase letters: alpha, beta, gamma, delta, epsilon, zeta, eta, theta, iota, kappa, lambda, mu, nu, xi, omicron, pi, rho, sigma, tau, upsilon, phi, chi, psi, omega

Arrows: rarr, larr, barr, uarr, darr, harr, rArr, lArr, uArr, dArr, hArr

Other Symbols: dagger, ddagger, deg, permil, pertenk, sect

See also

Examples

# Trademark symbol
symbol("My Companytrade")

# Registered Trademark symbol
symbol("My Companyreg")

# Dagger symbol concatenated
paste0(symbol("dagger"), "My footnotes")

# Alpha squared
symbol("alpha^[2]")

# Greek Symbols
symbol("SigmaPsiZeta")

# Useful Math Symbols
symbol("asymp ge le ne plusmn empty fnof radic sum")

# Useful Logical Symbols
symbol("forall isin notin cup cap and or")

# Chemistry
symbol("2H[2] + O[2] barr 2H[2]O")